requirements.txt
python-dotenv
pandas
numpy
torch
transformers
bitsandbytes
langchain-core
langchain-huggingface
tqdm
pydantic